refactor(incremental): move BuildStateCache into Calor.Compiler for CLI reuse
Shared by the MSBuild task and (next) the CLI compile path. Calor.Tasks,
its tests, and IncrementalBuildBench keep their unqualified names via
global using aliases; InternalsVisibleTo grants added. New shared API:
ComputeCliCompilerHash, ComputeOptionsHash(token), multi-directory
ComputeManifestHash, ComputeCommonDirectory, Delete.
